pyproject.toml
[tool.poetry]
name = "nonebot-plugin-playwright"
version = "0.2.0"
description = "NoneBot Playwright插件"
authors = ["eya46 <61458340+eya46@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "nonebot_plugin_playwright" }]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = ">=2.2.0"
playwright = "^1.0.0"
nonebot-plugin-localstore = { version = ">=0.7.1", optional = true }

[tool.poetry.extras]
localstore = ["nonebot-plugin-localstore"]

[tool.poetry.group.dev.dependencies]
nonebot-plugin-localstore = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 121
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E722"]